+++ backend-core/src/main/java/com/its/domovenok/core/dto/AuthInitResponse.java
@@ -1,0 +1,11 @@
+package com.its.domovenok.core.dto;
+
+import lombok.Value;
+
+@Value
+public class AuthInitResponse {
+
+    UserDto user;
+    String token;
+    boolean firstVisit;
+}