+++ backend-core/src/main/java/com/its/domovenok/core/dto/ErrorResponse.java
@@ -1,0 +1,9 @@
+package com.its.domovenok.core.dto;
+
+import lombok.Value;
+
+@Value
+public class ErrorResponse {
+
+    String error;
+}