+++ backend-core/src/main/java/com/its/domovenok/core/dto/GetPetsResponse.java
@@ -1,0 +1,10 @@
+package com.its.domovenok.core.dto;
+
+import java.util.List;
+import lombok.Value;
+
+@Value
+public class GetPetsResponse {
+
+    List<PetDto> pets;
+}